+++ abiwara-be-api/cmd/api/controller/visitor/controller_impl.go
@@ -72,7 +72,7 @@ func (controller *VisitorControllerImpl) FindAll(ctx *gin.Context) {
 		perPage,
 		querySearch,
 		uint(claims.RoleId),
-		uint(claims.Id),
+		claims.Id,
 	)
 	response.JsonPageData(ctx, http.StatusOK, "OK", visitorResponses, meta)
 }
